requirements.txt
absl-py==0.7.1
aiofiles==0.4.0
aiohttp==3.5.4
aioredis==1.2.0
alembic==1.0.11
APScheduler==3.6.1
asgiref==3.2.1
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
async-generator==1.10
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
autopep8==1.4.4
bcrypt==3.1.7
boto3==1.9.216
botocore==1.12.216
cached-property==1.5.1
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.14.0
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
daphne==2.3.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.0.2
django-filter==2.2.0
django-heroku==0.3.1
django-storages==1.7.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
fakeredis==1.0.4
fbmessenger==6.0.0
Flask==1.1.1
Flask-Cors==3.0.8
future==0.17.1
gast==0.2.2
gevent==1.4.0
greenlet==0.4.15
grpcio==1.23.0
gunicorn==19.9.0
h5py==2.9.0
hiredis==1.0.0
httpie==1.0.3
humanfriendly==4.18
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonpickle==1.2
jsonschema==2.6.0
kafka-python==1.4.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mattermostwrapper==2.1
mccabe==0.6.1
mock==3.0.5
msgpack==0.6.1
multidict==4.5.2
networkx==2.3
numpy==1.17.0
packaging==19.1
paramiko==2.6.0
pika==1.0.1
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.1
psycopg2==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pydot==1.4.1
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pykwalify==1.7.0
pylint==2.3.1
pymongo==3.9.0
PyNaCl==1.3.0
pyparsing==2.4.2
pypiwin32==223
pyreadline==2.1
PySocks==1.7.0
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-editor==1.0.4
python-telegram-bot==11.1.0
pytz==2019.2
pywin32==224
PyYAML==3.13
questionary==1.3.0
rasa-sdk==1.2.0
redis==3.3.8
requests==2.22.0
requests-toolbelt==0.9.1
rocketchat-API==0.6.34
ruamel.yaml==0.15.100
s3transfer==0.2.1
sanic-jwt==1.3.2
scikit-learn==0.20.4
scipy==1.3.1
simplejson==3.16.0
six==1.12.0
sklearn-crfsuite==0.3.6
slackclient==1.3.2
sortedcontainers==2.1.0
SQLAlchemy==1.3.7
sqlparse==0.3.0
tabulate==0.8.3
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminaltables==3.1.0
texttable==0.9.1
tqdm==4.35.0
twilio==6.29.4
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.24.3
wcwidth==0.1.7
webexteamssdk==1.1.1
websocket-client==0.54.0
websockets==6.0
Werkzeug==0.15.5
whitenoise==4.1.3
wincertstore==0.2
wrapt==1.11.2
yarl==1.3.0
zope.interface==4.6.0
